str_list: Add a well tested API for managing lists of strings.
Tested manually, with automated tests, as well as with valgrind.
With this str_list struct and supporting methods, we can easily
manage dynamic lists of strings.

Also in this commit we change our build process a bit, we compile
with the -g flag for being able to use GDB and we make "test" the
default make task.