st: Add symbol table models, creation, and deletion code.
Also added a default hash function (a polynomial rolling hash function)
and created a simple test. Valgrind shows no memory leaks and no risky
or invalid memory accesses.

We'll need to update the "Entry" structures later.

Also pending are the set and get functions. When writing these, I'll
need to provide an option to overwrite existing values for a given key.
I'll also need to write a helper function to see if a given key exists
or not.