st: Add a rough version of the symbol table generation code.
Some features and semantic checks are missing - for example,
module inputplist and outputplist are not updated in the symbol
table but they should.

I'm also pretty sure that there are a bugs that need to be squashed.
As of yet valgrind shows no invalid memory references, and the code
runs without segfaults, so it can be published as rough "alpha" code.

Also a bug in dynamic_range in the AST code was fixed (forgot to
traverse children first) and a reference from leaf node to symbol
table node was added.